requirements.txt
fastapi[standard]
python-dotenv
polars
ngrok
pydantic
pyyaml
rapidfuzz
rq
sqlalchemy
psycopg2-binary
pandas